timeout working

+++ testing/ZIPsFS_try_install_from_github.sh
@@ -1,15 +1,18 @@
 #!/usr/bin/env bash
 set -u
 
+
 readonly DIR=~/test/ZIPsFS/installation
 readonly U=https://github.com/christophgil/ZIPsFS/archive/refs/heads/main.zip
 
 main(){
-    mkdir -p $DIR || return
-    cd $DIR || return
+    rm -r ~/test/ZIPsFS/installation
+    [[ -d $DIR ]] && return 1
+    mkdir -p $DIR || return 1
+    cd $DIR || return 1
     rm -r ZIPsFS-main main.zip 2>/dev/null
-    wget -N $U || return
-    unzip -o main.zip || return
+    wget -N $U || return 1
+    unzip -o main.zip || return 1
     ls -l
     ./ZIPsFS-main/src/ZIPsFS.compile.sh
 }
